MidReview/IR_Project_BE/requirements.txt
attrs==22.2.0
beautifulsoup4==4.12.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.7
Fiona==1.9.2
Flask==2.2.3
Flask-Cors==3.0.10
geopandas==0.12.2
huggingface-hub==0.13.4
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
langcodes==3.3.0
locationtagger==0.0.1
lxml==4.9.2
MarkupSafe==2.1.2
munch==2.5.0
murmurhash==1.0.9
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pycountry==22.3.5
pydantic==1.10.7
pyproj==3.5.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
snscrape==0.6.2.20230320
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
